Fix bug which prevented repo from rendering
There is still an problem with the events not binding properly (or being
unbound) on the initial render of the raw view. This causes the clicking
on the table not to behave correctly, ie. sorting and item details don't
work on click. If you switch tabs then it works fine, so it's something
about the initial event state.

Also fixes a bug which prevents charts from being created. Charts themselves
are still broken (always been though).